Add files via upload
LFA Interface.  Recognition Rate: Approximately 97%.
Supported Image Types: This program is only compatible with chest radiographic images. Ensure you use the appropriate type of X-ray.
Recommended Minimum Resolution: It is advised not to use images with a resolution lower than 256x256.
Important Considerations:

If the probability of having COVID-19 is in the range of 50% to 60%, it is recommended to conduct additional tests on the patient to rule out the disease.
This program is specifically designed for the detection of COVID-19. However, in cases where pneumonia may be present, the program might classify it as COVID-19, reducing the recognition rate to 90%. Therefore, caution is advised in such situations.

*Note: The information provided by this program does not replace the clinical judgment of a healthcare professional and should be used as a complementary tool in the diagnostic process.